+++ rust/operator-binary/src/config/mod.rs
@@ -13,7 +13,7 @@ use stackable_hdfs_crd::{HdfsCluster, HdfsPodRef};
 use stackable_operator::utils::cluster_info::KubernetesClusterInfo;
 use std::collections::BTreeMap;
 
-pub mod jvm_arguments;
+pub mod jvm;
 
 #[derive(Clone)]
 pub struct HdfsSiteConfigBuilder {
